+++ demo/latest/src/App.tsx
@@ -2,12 +2,13 @@ import React from 'react';
 import Viewer, { Worker } from '@phuocng/react-pdf-viewer';
 
 import '@phuocng/react-pdf-viewer/cjs/react-pdf-viewer.css';
+import RenderErrorExample from './RenderErrorExample';
 
 const App = () => {
     return (
         <Worker workerUrl="https://unpkg.com/pdfjs-dist@2.2.228/build/pdf.worker.min.js">
             <div style={{ height: '750px', padding: '16px 0' }}>
-                <Viewer fileUrl="/pdf-open-parameters.pdf" />
+                <RenderErrorExample fileUrl="/pdf-open-parameters1.pdf" />
             </div>
         </Worker>
     );